+++ apps/global-registry/src/hooks/useRegistryStore.ts
@@ -1,0 +1,7 @@
+import { useAppStore } from './useAppStore';
+
+export const useRegistryStore = () => {
+  const appStore = useAppStore();
+
+  return appStore.registryStore;
+};