+++ packages/api/src/index.ts
@@ -2,3 +2,4 @@ export * from './IndexerApi';
 export * from './FaucetApi';
 export * from './StatsApi';
 export * from './ClusterManagementApi';
+export * from './AccessRegistryApi';